move tests to test_transforming
